requirements.txt
PyAudio==0.2.14
PyGObject==3.50.0
pycairo==1.27.0
vosk==0.3.45
evdev==1.7.1
numpy==2.2.1
cffi==1.17.1
pycparser==2.22
six==1.17.0
dbus-python==1.3.2